Just tidying up tests slightly
